# src/main/java/org/servament/dto/UpdateBookingDTO.java
package org.servament.dto;

import java.util.UUID;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UpdateBookingDTO {

    @NotNull
    private UUID consumer;
}