# src/main/resources/apis/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Event-Ease API",
    "description" : "Seamlessly manage any event quickly and easily with our versatile SaaS platform. Perfect for any occasion.",
    "termsOfService" : "Your terms here",
    "contact" : {
      "name" : "Example API Support",
      "url" : "http://exampleurl.com/contact",
      "email" : "coralo.luca@outlook.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/license/mit"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
  "paths" : {
    "/booking" : {
      "post" : {
        "tags" : [ "Booking Resource" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateBookingDTO"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookingDTO"
                }
              }
            }
          }
        }
      }
    },
    "/bookings" : {
      "get" : {
        "tags" : [ "Booking Resource" ],
        "parameters" : [ {
          "name" : "events",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "statuses",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BookingStatus"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BookingDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bookings/{id}" : {
      "get" : {
        "tags" : [ "Booking Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookingDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Booking Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "head" : {
        "tags" : [ "Booking Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "Booking Resource" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateBookingDTO"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookingDTO"
                }
              }
            }
          }
        }
      }
    },
    "/bookings:paged" : {
      "get" : {
        "tags" : [ "Booking Resource" ],
        "parameters" : [ {
          "name" : "events",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "format" : "uuid",
              "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
              "type" : "string"
            }
          }
        }, {
          "name" : "numPage",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "name" : "statuses",
          "in" : "query",
          "schema" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BookingStatus"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginationBookingDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "BookingDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "event" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "eventCode" : {
            "type" : "string"
          },
          "supplier" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "consumer" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "status" : {
            "$ref" : "#/components/schemas/BookingStatus"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "BookingStatus" : {
        "enum" : [ "PENDING", "CONFIRMED", "REJECTED", "CANCELLED" ],
        "type" : "string"
      },
...
      "CreateBookingDTO" : {
        "required" : [ "event", "consumer" ],
        "type" : "object",
        "properties" : {
          "event" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "consumer" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }
      },
...
      "Instant" : {
        "format" : "date-time",
        "type" : "string",
        "example" : "2022-03-10T16:15:50Z"
      },
...
      "PaginationBookingDTO" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BookingDTO"
            }
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalSize" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
...
      "UUID" : {
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}",
        "type" : "string"
      },
      "UpdateBookingDTO" : {
        "required" : [ "consumer" ],
        "type" : "object",
        "properties" : {
          "consumer" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }
      },
...
    }
  }
}